restart preview after picture taken
0.6.1 - A new property has been added (pictureTimeout) that will
restart the preview window x number of milliseconds after an image is
taken. The default is 1000ms (1 second). Passing in a negative number
causes the preview not to restart.
Other minor changes include updates to documentation and example app.js
to reflect this, the latest dist, and some code clean up.